% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_metagate}
\alias{run_metagate}
\title{Run shiny app}
\usage{
run_metagate(
  debug = getOption("metagate.debug", FALSE),
  modules = getOption("metagate.modules"),
  populations = getOption("metagate.populations", 256),
  groups = getOption("metagate.groups", 20),
  asterisk_limits = getOption("metagate.asterisk_limits", c(0.05, 0.01, 0.001, 0.0001)),
  posthoc_limit = getOption("metagate.posthoc_limit", 0.05),
  port = NULL,
  file = getOption("metagate.file"),
  disable = getOption("metagate.disable")
)
}
\arguments{
\item{debug}{Run metagate in debug mode. If \code{TRUE}, more information will be shown in the R console, and raw data
can be accessed in the "Debug" menu inside MetaGate. Defaults to \code{getOption("metagate.debug")}, if set, or \code{FALSE}
if not.}

\item{modules}{Character vector with names of custom analysis modules that should be loaded. Defaults to
\code{getOption("metagate.modules")}.}

\item{populations}{The number of populations that can be defined when importing data. Setting a high number may
impact performance. Defaults to \code{getOption("metagate.populations")}, if set, or \code{256} if not.}

\item{groups}{The number of groups that can be defined during analysis. Setting a high number may impact
performance. Defaults to \code{getOption("metagate.groups")}, if set, or \code{20} if not.}

\item{asterisk_limits}{Numerical vector of cut-offs for displaying p values as asterisks. A p value will be shown as
n asterisks if below or equal to the n'th value of this vector. Defaults to
\code{getOption("metagate.asterisk_limits")}, if set, or \code{c(0.05, 0.01, 0.001, 0.0001)} if not.}

\item{posthoc_limit}{In statistical comparisons of more than two groups, post-hoc testing will only be performed if
initial analysis of variance yields a p value below or equal to this limit. Defaults to
\code{getOption("metagate.posthoc_limit")}, or \code{0.05} if not.
if set, or \code{0.05} if not.}

\item{port}{Shiny web server port. If not set, a random port is used.}

\item{file}{MetaGate file to load automatically upon launch. Defaults to getOption("metagate.file").}

\item{disable}{Character vector with names of features that should be disabled. Defaults to
getOption("metagate.disable"). Could include the following:
\itemize{
\item \code{open_saved}: User will not be allowed to load saved projects.
\item \code{import_flowjo}: Creating new projects based on FlowJo workspaces will not be allowed.
\item \code{import_cytobank}: Creating new projects based on Cytobank gating files will not be allowed.
\item \code{merge_projects}: Disable project merging feature.
\item \code{save_project}: User will not be allowed to save projects as MetaGate files.
\item \code{close_project}: User will not be allowed to close the current project.
}}
}
\description{
This function launches the graphical user interface in the web browser.
}
\examples{
\dontrun{
# Launch MetaGate with default settings
run_metagate()

# Automatically load a project upon MetaGate launch, and allow only analysis 
# within this project:
run_metagate(file = "/path/to/my/metagate/file.metagate",
             disable = c("open_saved", "import_flowjo", "import_cytobank",
                         "merge_projects", "save_project", "close_project"))
}
}
\keyword{metagate}
